% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModulonCore_functions.R
\name{build.graph}
\alias{build.graph}
\title{Build the graph}
\usage{
build.graph(net, weight = WEIGHT)
}
\arguments{
\item{net}{A dataframe with a network encoded in 3 columns: 'Source','Interaction','Target'.}

\item{weight}{A vector with the weight of each interaction. If none is provided, it will be assumed the same weight (1) for all the interactions}
}
\value{
A dataframe with the network encoded in 3 columns:  'Source','Target','Weight'.
}
\description{
Find all regulatory regulatory events between TFs and add weights for the regulatory interactions
}
\details{
The resulting network only contains regulatory events between TFs
}
\examples{
\dontrun{
if(interactive()){
 build.graph(network.TILs)
 }
}
}
