% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-ModulonCore_functions.R
\name{build.igraph}
\alias{build.igraph}
\title{FUNCTION_TITLE}
\usage{
build.igraph(net)
}
\arguments{
\item{net}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[igraph]{graph_from_edgelist}}, \code{\link[igraph]{E}}
}
