% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-ModulonCore_functions.R
\name{connectivity}
\alias{connectivity}
\title{FUNCTION_TITLE}
\usage{
connectivity(
  net,
  cc,
  min.size,
  dist.method,
  nperm,
  vertex.attr,
  edge.attr,
  significance
)
}
\arguments{
\item{net}{PARAM_DESCRIPTION}

\item{cc}{PARAM_DESCRIPTION}

\item{min.size}{PARAM_DESCRIPTION}

\item{dist.method}{PARAM_DESCRIPTION}

\item{nperm}{PARAM_DESCRIPTION}

\item{vertex.attr}{PARAM_DESCRIPTION}

\item{edge.attr}{PARAM_DESCRIPTION}

\item{significance}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[igraph]{vertex_attr}}, \code{\link[igraph]{V}}
\code{\link[SANTA]{Knet}}
}
