% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModulonCore_functions.R
\name{core.out}
\alias{core.out}
\title{Print out regulatory core analysis results}
\usage{
core.out(net, cc, regulatory.cores, dir = "./")
}
\arguments{
\item{net}{Network built with build.graph()}

\item{cc}{Connected components generated with find.connected.components()}

\item{regulatory.cores}{Regulatory cores as the output of core()}

\item{dir}{Output directory}
}
\value{
Files with regulatory core networks and constituent elements printed out to the output directory
}
\description{
Generate files with regulatory core networks and constituent elements.
}
\details{
The name of each file contains the corresponding modulon and connected component
}
\examples{
\dontrun{
if(interactive()){
 #'  graph = build.graph(network.TILs)
 igraph = build.igraph(graph)
 cc=find.connected.components(net=igraph,modulons = modulons.TILs)
 results.connectivity=connectivity( net=igraph, cc=cc, min.size='5',
 dist.method="shortest.paths",
 nperm=100,
 vertex.attr="query",
 edge.attr="Weight",
 significance=0.05)
 regulatory.cores=core(results.connectivity=results.connectivity)
 core.out(net=graph,cc=cc,regulatory.cores=regulatory.cores,dir="./cores")
 }
}
}
