% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModulonCore_functions.R
\name{core}
\alias{core}
\title{Regulatory core identification}
\usage{
core(results.connectivity, significance = 0.05)
}
\arguments{
\item{results.connectivity}{A list with the output generated by the function connectivity().}
}
\value{
A list with all the regulatory cores of each modulon
}
\description{
Evaluate the results of the connectivity analysis and collect connected components with statistically supported over-connectivity
}
\details{
Modulons may have none, one or more than one regulatory cores
}
