% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModulonCore_functions.R,
%   R/oxy-ModulonCore_functions.R
\name{find.connected.components}
\alias{find.connected.components}
\title{Connected Components Identification}
\usage{
find.connected.components(net, modulons, mode = MODE, dir = DIR)

find.connected.components(net, modulons, mode = MODE, dir = DIR)
}
\arguments{
\item{net}{PARAM_DESCRIPTION}

\item{modulons}{PARAM_DESCRIPTION}

\item{mode}{PARAM_DESCRIPTION, Default: MODE}

\item{dir}{PARAM_DESCRIPTION, Default: DIR}
}
\value{
A list with as many elements as modulons/clusters containing  all the connected components

OUTPUT_DESCRIPTION
}
\description{
Find all connected components within modulons or any cluster of genes

FUNCTION_DESCRIPTION
}
\details{
The names of connected components are encoded as ccX, being 'X' the ID number of the connected component

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 graph = build.graph(network.TILs)
 igraph = build.igraph(graph)
 find.connected.components(net=igraph,modulons = modulons.TILs)
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[igraph]{subgraph}}, \code{\link[igraph]{V}}, \code{\link[igraph]{component_distribution}}
}
