% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{network.TILs}
\alias{network.TILs}
\title{network.TILs Data}
\format{
A dataframe encoding the network: \describe{ \item{x}{a dataframe of dimension
\code{12490 x 3} with the following colums:\describe{\item{Source}{Transcription factor} \item{Interaction}{Regulatory interaction} \item{Target}{Target gene regulated by the transcription factor}
}
\source{
\url{https://www.nature.com/articles/s41467-021-23324-4}
}
\usage{
network.TILs
}
\description{
A dataset containing the transcriptional network inferred using SCENIC pipeline on murine TILs transcriptomics.
}
\examples{
network.TILs
}
\keyword{datasets}
