% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModulonCore_functions.R,
%   R/oxy-ModulonCore_functions.R
\name{range01}
\alias{range01}
\title{Range between 0 and 1}
\usage{
range01(x)

range01(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
Normalized values (x-min(x))/(max(x)-min(x))

OUTPUT_DESCRIPTION
}
\description{
Normalize absolute values of a numeric vector

FUNCTION_DESCRIPTION
}
\details{
DETAILS

DETAILS
}
\examples{
\dontrun{
if(interactive()){
range01(c('-50','-5','5','50','100')
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
