% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModulonCore_functions.R,
%   R/oxy-ModulonCore_functions.R
\name{redundancy}
\alias{redundancy}
\title{Regulon Redundancy
Calculate redundancy between two regulons}
\usage{
redundancy(a, b)

redundancy(a, b)
}
\arguments{
\item{a}{PARAM_DESCRIPTION}

\item{b}{PARAM_DESCRIPTION}
}
\value{
Redundancy calculated as the ratio between the intersection and the minimum length of the two vectors.

OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS

DETAILS
}
\examples{
\dontrun{
if(interactive()){
redundancy(c('A','B','C','D','E'), c('A','B','C'))
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
