% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModulonCore_functions.R
\name{tab.cc}
\alias{tab.cc}
\title{Connected components table generation}
\usage{
tab.cc(cc, cores = NULL)
}
\arguments{
\item{cc}{A list with all the connected components of each modulon.}
}
\value{
A table of 4 columns
}
\description{
Generate a table with all the connected components and their constituent elements
}
\examples{
\dontrun{
if(interactive()){
 cc = cc.TILs
 tab.cc(cc)
 }
}
}
