% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModulonCore_functions.R,
%   R/oxy-ModulonCore_functions.R
\name{build.graph}
\alias{build.graph}
\alias{build.igraph}
\title{Build the graph}
\usage{
build.graph(net, weight = WEIGHT)

build.igraph(net)

build.graph(net, weight = WEIGHT)
}
\arguments{
\item{net}{PARAM_DESCRIPTION}

\item{weight}{PARAM_DESCRIPTION, Default: WEIGHT}
}
\value{
A dataframe with the network encoded in 3 columns:  'Source','Target','Weight'.

A network as an igraph object

OUTPUT_DESCRIPTION
}
\description{
Find all regulatory regulatory events between TFs and add weights for the regulatory interactions

Covert the input network into an igraph object

FUNCTION_DESCRIPTION
}
\details{
The resulting network only contains regulatory events between TFs

The resulting network only contains regulatory events between TFs

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 build.graph(network.TILs)
 }
}
\dontrun{
if(interactive()){
 graph = build.graph(network.TILs)
 build.igraph(graph)
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
