% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModulonCore_functions.R
\name{build.igraph}
\alias{build.igraph}
\title{Build the igraph}
\usage{
build.igraph(net)
}
\arguments{
\item{net}{A dataframe with the network encoded in 3 columns:  'Source','Target','Weight'.}
}
\value{
A network as an igraph object
}
\description{
Covert the input network into an igraph object
}
\details{
The resulting network only contains regulatory events between TFs
}
\examples{
\dontrun{
if(interactive()){
 graph = build.graph(network.TILs)
 build.igraph(graph)
 }
}
}
