% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModulonCore_functions.R
\name{connectivity}
\alias{connectivity}
\title{SANTA analysis}
\usage{
connectivity(
  net,
  cc,
  min.size,
  dist.method,
  nperm,
  vertex.attr,
  edge.attr,
  significance
)
}
\arguments{
\item{cc}{A list with all the connected components of each modulon}
}
\value{
A list with as many elements as modulons containing the results of the SANTA analysis
}
\description{
Evaluate the statistical support for connected components over-connectivity
}
\details{
The results can be plotted as follows: plot(x[[][[]]])
}
\examples{
\dontrun{
if(interactive()){
 graph = build.graph(network.TILs)
 igraph = build.igraph(graph)
 cc=find.connected.components(net=igraph,modulons = modulons.TILs)
 connectivity( cc=cc, min.size='5',
 dist.method="shortest.paths",
 nperm=100,
 vertex.attr="query",
 edge.attr="Weight",
 significance=0.05)
 }
}
}
