% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-ModulonCore_functions.R
\name{core.out}
\alias{core.out}
\title{FUNCTION_TITLE}
\usage{
core.out(net, cc, regulatory.cores, dir = "./")
}
\arguments{
\item{net}{PARAM_DESCRIPTION}

\item{cc}{PARAM_DESCRIPTION}

\item{regulatory.cores}{PARAM_DESCRIPTION}

\item{dir}{PARAM_DESCRIPTION, Default: './'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[limma]{strsplit2}}
}
