% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-ModulonCore_functions.R
\name{core}
\alias{core}
\title{FUNCTION_TITLE}
\usage{
core(results.connectivity, significance = 0.05)
}
\arguments{
\item{results.connectivity}{PARAM_DESCRIPTION}

\item{significance}{PARAM_DESCRIPTION, Default: 0.05}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
