% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModulonCore_functions.R
\name{find.connected.components}
\alias{find.connected.components}
\title{Connected Components Identification}
\usage{
find.connected.components(net, modulons, mode = MODE, dir = DIR)
}
\arguments{
\item{net}{A network as an igraph object}

\item{modulons}{A list with as many elements as modulons/clusters containing the constituent elements.}
}
\value{
A list with as many elements as modulons/clusters containing  all the connected components
}
\description{
Find all connected components within modulons or any cluster of genes
}
\details{
The names of connected components are encoded as ccX, being 'X' the ID number of the connected component
}
\examples{
\dontrun{
if(interactive()){
 graph = build.graph(network.TILs)
 igraph = build.igraph(graph)
 find.connected.components(net=igraph,modulons = modulons.TILs)
 }
}
}
