% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{jaccard}
\alias{jaccard}
\title{Jaccard Distance}
\usage{
jaccard(a, b)
}
\arguments{
\item{b}{A vector.}
}
\value{
Jaccard distance calculated as the ratio between the intersection and the union of the two vectors.
}
\description{
Calculate Jaccard distance between two vectors
#' @param a A vector.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 jaccard(c('A','B','C','D','E'), c('A','B'))
 }
}
}
