% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModulonCore_functions.R,
%   R/oxy-ModulonCore_functions.R
\name{jaccard}
\alias{jaccard}
\title{Jaccard Distance}
\usage{
jaccard(a, b)

jaccard(a, b)
}
\arguments{
\item{a}{PARAM_DESCRIPTION}

\item{b}{PARAM_DESCRIPTION}
}
\value{
Jaccard distance calculated as the ratio between the intersection and the union of the two vectors.

OUTPUT_DESCRIPTION
}
\description{
Calculate Jaccard distance between two vectors

FUNCTION_DESCRIPTION
}
\details{
DETAILS

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 jaccard(c('A','B','C','D','E'), c('A','B','C'))
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
