% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModulonCore_functions.R, R/functions.R
\name{find.connected.components}
\alias{find.connected.components}
\alias{connectivity}
\alias{core}
\alias{print.core}
\alias{jaccard}
\title{Connected Components Identification}
\usage{
find.connected.components(net, modulons, mode = MODE, dir = DIR)

connectivity(
  cc,
  min.size = MINSIZE,
  dist.method,
  nperm,
  vertex.attr,
  edge.attr,
  significance
)

core(modulon.subnetworks.connected.components.SANTA, significance = 0.05)

\method{print}{core}(net, SANTA.INPUT, regulatory.cores, dir = DIR)

jaccard(a, b)
}
\arguments{
\item{net}{A list with all the connected components of each modulon}

\item{modulons}{A list.}

\item{cc}{A list with all the connected components of each modulon}

\item{a}{A vector.
#' @param a A vector.}

\item{b}{A vector.}
}
\value{
A list with as many elements as modulons containing the constituent transcripton factors of all the connected components of each modulon

A list with as many elements as modulons containing the constituent transcripton factors of all the connected components of each modulon

A list with as many elements as modulons containing the constituent transcripton factors of all the connected components of each modulon

A list with as many elements as modulons containing the constituent transcripton factors of all the connected components of each modulon

Jaccard distance calculated as the ratio between the intersection and the union of the two vectors.
}
\description{
Find all connected components within in modulon

Find all regulatory cores within in modulon

Find all regulatory cores within in modulon

Generate files with connected component networks and connected component elements

Calculate Jaccard distance between two vectors
}
\details{
The names of the outpuT list are composed by....

The names of the outpuT list are composed by....

The names of the outpuT list are composed by....

The names of the outpuT list are composed by....

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 jaccard(c('A','B','C','D','E'), c('A','B','C'))
 }
}
\dontrun{
if(interactive()){
 jaccard(c('A','B','C','D','E'), c('A','B','C'))
 }
}
\dontrun{
if(interactive()){
 jaccard(c('A','B','C','D','E'), c('A','B','C'))
 }
}
\dontrun{
if(interactive()){
 jaccard(c('A','B','C','D','E'), c('A','B','C'))
 }
}
\dontrun{
if(interactive()){
 jaccard(c('A','B','C','D','E'), c('A','B','C'))
 }
}
}
