% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModulonCore-package.R
\docType{package}
\name{ModulonCore-package}
\alias{ModulonCore}
\alias{ModulonCore-package}
\title{ModulonCore: Modulon Connected Components Identification}
\description{
ModulonCore finds all connected components within the modulon regulatory network.
}
\author{
\strong{Maintainer}: Isaac Crespo \email{isaaccrespo@hotmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
