% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{range01}
\alias{range01}
<<<<<<< HEAD
\title{Range between 0 and 1}
=======
\title{Range \link{0,1}}
>>>>>>> 661fb783d47c9768f2fac50309ab297c738490cd
\usage{
range01(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
Normalized values (x-min(x))/(max(x)-min(x))
}
\description{
Normalize absolute values of a numeric vector
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
range01(c('-50','-5','5','50','100')
 }
}
}
