% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-functions2.R
\name{redundancy3}
\alias{redundancy3}
\title{FUNCTION_TITLE}
\usage{
redundancy3(a, b)
}
\arguments{
\item{a}{PARAM_DESCRIPTION}

\item{b}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 redundancy3(c('A','B','C','D','E'), c('A','B','C'))
 }
}
}
